moved preferences form into component folder
